good old Database "abstraction layer"
this so that we get to see some stuff running quickly. I know that this
fixes phexrep to only work with PostgreSQL, but I'll leave "needs to
work with any database" as an open ticket for later.

I need PostgreSQL support and I know exactly how this Database class
works (I wrote it in 2004), so this is what I'm using now.