report details
the original Rails plugin is still showing the list of exceptions below
the current report. I'm recreating this here.

But now, whenever I'm rendering a detail directly without having first
rendered the index view, the list will be missing, so we'll have to
fetch it.

This means that I have to extract the list-rendering functionality into
its own function which leads to some indirectiony with $.sammy(), so
thich patch looks like it's mixing whitespace with code changes which it
kinda isn't - the indentation change come with the code change